+++ src/main/java/hu/eszter/bokkon/model/animal/BigDog.java
@@ -12,9 +12,9 @@ public final class BigDog implements ProtectorAnimal {
     }
 
     @Override
-    public Map<Animal, Integer> changeableTo() {
-        Map<Animal, Integer> result = new HashMap<>();
-        result.put(new Cow(), 1);
+    public Map<Animal, Double> changeableTo() {
+        Map<Animal, Double> result = new HashMap<>();
+        result.put(new Cow(), 1.0);
         return result;
     }
